move release scripts to folder
