+++ script/release/prepare-release.js
@@ -11,8 +11,8 @@ const { GitProcess } = require('dugite')
 
 const path = require('path')
 const readline = require('readline')
-const releaseNotesGenerator = require('./release-notes/index.js')
-const { getCurrentBranch } = require('./lib/utils.js')
+const releaseNotesGenerator = require('../release-notes/index.js')
+const { getCurrentBranch } = require('../lib/utils.js')
 const bumpType = args._[0]
 const targetRepo = bumpType === 'nightly' ? 'nightlies' : 'electron'
