+++ script/release/release.js
@@ -15,7 +15,7 @@ const fs = require('fs')
 const { execSync } = require('child_process')
 const nugget = require('nugget')
 const got = require('got')
-const pkg = require('../package.json')
+const pkg = require('../../package.json')
 const pkgVersion = `v${pkg.version}`
 const pass = '\u2713'.green
 const path = require('path')
