+++ script/release/upload-symbols.py
